+++ components/InputBox/InputBox.module.scss
@@ -3,7 +3,7 @@
   font-size: 16px;
   padding: 18px 15px;
   border-radius: 8px;
-  background: var(--Linkbrary-white, #fff);
+  background: #fff;
 }
 
 .input-wrapper input:focus {
@@ -30,7 +30,6 @@
   border-radius: 8px;
   color: #f5f5f5;
   font-size: 18px;
-  font-family: Pretendard;
   font-weight: 600;
   margin-top: 6px;
 }
@@ -46,5 +45,5 @@
 }
 
 .default-border {
-  border: 1px solid var(--Linkbrary-gray20, #ccd5e3);
+  border: 1px solid #ccd5e3;
 }
