+++ components/SignInForm/SingInForm.module.scss
@@ -13,7 +13,6 @@
   border-radius: 8px;
   color: #f5f5f5;
   font-size: 18px;
-  font-family: Pretendard;
   font-weight: 600;
   margin-top: 6px;
 }
