+++ components/SocialLogin/SocialLogin.module.scss
@@ -33,7 +33,6 @@
 .social-login p {
   color: #373740;
   font-size: 14px;
-  font-family: Pretendard;
   font-weight: 400;
 }
