+++ components/Sorting/Sorting.module.scss
@@ -11,14 +11,13 @@
 
 .button-wrapper button {
   border-radius: 5px;
-  border: 1px solid var(--Linkbrary-primary-color, #6d6afe);
+  border: 1px solid #6d6afe;
   background: #fff;
   padding: 8px 12px;
 }
 
 .sort-name {
   color: #000;
-  font-family: Pretendard;
   font-size: 16px;
   font-style: normal;
   font-weight: 400;
@@ -26,8 +25,6 @@
 
 .folder-name {
   color: #000;
-  /* Linkbrary/h3-semibold */
-  font-family: Pretendard;
   font-size: 24px;
   font-style: normal;
   font-weight: 600;
