+++ components/index.js
@@ -8,3 +8,7 @@ export * from "./SearchBar";
 export * from "./Sorting";
 export * from "./ModalLayout";
 export * from "./FolderTitle";
+export * from "./SignUpForm";
+export * from "./SignInForm";
+export * from "./SocialLogin";
+export * from "./InputBox";
