+++ pages/_app.tsx
@@ -1,5 +1,5 @@
 import type { AppProps } from "next/app";
-import "styles/giobal.css";
+import "styles/global.css";
 
 export default function App({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />;
