+++ pages/signin.js
@@ -1,8 +1,7 @@
 import Link from "next/link";
 import styles from "../styles/sign.module.scss";
 import classNames from "classnames/bind";
-import { SinginForm } from "../components/SinginForm/SinginForm";
-import { SocalLogin } from "../components/SocalLogin/SocalLogin";
+import { SignInForm, SocialLogin } from "../components";
 
 const cx = classNames.bind(styles);
 
@@ -16,13 +15,13 @@ function singInPage() {
           </Link>
           <h2>
             회원이 아니신가요?
-            <Link className={cx("singin-link")} href="./signup">
+            <Link className={cx("singIn-link")} href="./signup">
               회원가입하기
             </Link>
           </h2>
         </div>
-        <SinginForm />
-        <SocalLogin />
+        <SignInForm />
+        <SocialLogin />
       </div>
     </div>
   );
