+++ styles/sign.module.scss
@@ -27,7 +27,6 @@
 .title h2 {
   color: black;
   font-size: 16px;
-  font-family: Pretendard;
   font-weight: 400;
   line-height: 24px;
 }
@@ -35,11 +34,10 @@
 .title a {
   color: #6d6afe;
   font-size: 16px;
-  font-family: Pretendard;
   font-weight: 600;
 }
 
-.singin-link {
+.signIn-link {
   margin-left: 8px;
 }
