+++ utils/url.js
@@ -4,7 +4,7 @@ export const ApiUrl = {
   sampleUser: `${baseUrl}sample/user`,
   usersFolders: `${baseUrl}users/1/folders`,
   usersLinks: `${baseUrl}users/1/links`,
-  signin: `${baseUrl}sign-in`,
-  signup: `${baseUrl}sign-up`,
+  signIn: `${baseUrl}sign-in`,
+  signUp: `${baseUrl}sign-up`,
   checkEmail: `${baseUrl}check-email`,
 };
